+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## (Unreleased)
 
 ### Monument
+- (#91) Calls can now go from/to different lead labels.  Set this with e.g.
+    `lead_location = { from = "2nds", to = "HL" }`.  Useful for adding finer control over where
+    calls can be placed.
 - (#91) Allow multiple labels on the same row within a lead.  Also reversed the syntax from e.g.
     `lead_locations = { 0 = "LE", 16 = "HL" }` to `lead_locations = { LE = 0, HL = 16 }`.  The same
     label can be added to multiple rows like `lead_locations = { SE = [3, 9] }` (for Six-Ends in